# src/main/java/ru/opencode/bankinfo/messages/dto/subDTO/EMessageNameDTO.java
package ru.opencode.bankinfo.messages.dto.subDTO;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import lombok.Data;

@Data
public class EMessageNameDTO {
    @NotNull
    @NotBlank
    private String name;
}